% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution par semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant une colonne 'Jour de la semaine'.}
}
\value{
Un data.frame avec le total par jour.
}
\description{
Groupe et additionne les trajets par jour de la semaine.
}
