% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant une colonne 'Numéro de boucle'.}
}
\value{
Le nombre de boucles distinctes.
}
\description{
Calcule le nombre de boucles de comptage uniques.
}
