% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Données de comptage de vélos à Nantes (Toussaint 2025)}
\format{
Un data.frame avec les colonnes suivantes :
\describe{
\item{Numéro de boucle}{Identifiant unique de la boucle}
\item{Jour}{Date du comptage}
\item{Total}{Nombre total de passages sur la journée}
\item{Jour de la semaine}{Identifiant du jour (1 pour lundi, etc.)}
\item{Probabilité de présence d'anomalies}{Indicateur de fiabilité des données}
\item{Boucle de comptage}{Nom complet et direction de la boucle}
}
}
\source{
\url{https://data.nantesmetropole.fr/}
}
\usage{
df_velo
}
\description{
Un jeu de données contenant les comptages horaires des boucles de
comptage de Nantes Métropole durant les vacances de la Toussaint 2025.
}
\keyword{datasets}
