% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les données vélos.}
}
\value{
Un data.frame filtré.
}
\description{
Conserve uniquement les trajets où aucune anomalie n'est détectée.
}
