% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(trajet, boucle = NULL)
}
\arguments{
\item{trajet}{Un data.frame contenant les données vélos.}

\item{boucle}{Un vecteur contenant les numéros de boucle à garder. Si NULL, ne filtre pas.}
}
\value{
Un data.frame filtré.
}
\description{
Conserve uniquement les données correspondant aux boucles spécifiées.
}
