% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Tracer la distribution par semaine}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame de trajets bruts.}
}
\value{
Un graphique ggplot.
}
\description{
Génère un graphique en barres de la fréquentation par jour.
}
