% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet maximum}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les colonnes 'Total', 'Jour' et 'Boucle de comptage'.}
}
\value{
Un data.frame d'une ligne avec le record.
}
\description{
Isole la ligne comportant le plus grand nombre de trajets.
}
